Flyttet handlers til egen fil
